+++ __init__.py
@@ -4,22 +4,40 @@ from .taphome_sdk import *
 from homeassistant.helpers.discovery import async_load_platform
 import voluptuous
 import homeassistant.helpers.config_validation as config_validation
-from homeassistant.const import CONF_TOKEN, CONF_LIGHTS, CONF_COVERS, CONF_SWITCHES
+
+from homeassistant.const import (
+    CONF_TOKEN,
+    CONF_LIGHTS,
+    CONF_COVERS,
+    CONF_SWITCHES,
+    CONF_SENSORS,
+    CONF_BINARY_SENSORS,
+)
 
 DOMAIN = "taphome"
 TAPHOME_API_SERVICE = f"{DOMAIN}_TapHomeApiService"
 TAPHOME_DEVICES = f"{DOMAIN}_Devices"
-TAPHOME_CORES = "cores"
+TAPHOME_LANGUAGE = f"{DOMAIN}_language"
+CONF_CORES = "cores"
+CONF_LANGUAGE = "language"
 CONF_CLIMATES = "climates"
 
 CONFIG_SCHEMA = voluptuous.Schema(
     {
         DOMAIN: voluptuous.Schema(
             {
-                TAPHOME_CORES: [
+                voluptuous.Optional(
+                    CONF_LANGUAGE, default="en"
+                ): config_validation.string,
+                CONF_CORES: [
                     voluptuous.All(
                         config_validation.has_at_least_one_key(
-                            CONF_LIGHTS, CONF_COVERS, CONF_CLIMATES, CONF_SWITCHES
+                            CONF_LIGHTS,
+                            CONF_COVERS,
+                            CONF_CLIMATES,
+                            CONF_SWITCHES,
+                            CONF_SENSORS,
+                            CONF_BINARY_SENSORS,
                         ),
                         {
                             voluptuous.Required(CONF_TOKEN): config_validation.string,
@@ -35,9 +53,15 @@ CONFIG_SCHEMA = voluptuous.Schema(
                             voluptuous.Optional(
                                 CONF_SWITCHES, default=[]
                             ): config_validation.ensure_list,
+                            voluptuous.Optional(
+                                CONF_SENSORS, default=[]
+                            ): config_validation.ensure_list,
+                            voluptuous.Optional(
+                                CONF_BINARY_SENSORS, default=[]
+                            ): config_validation.ensure_list,
                         },
                     )
-                ]
+                ],
             }
         )
     },
@@ -46,7 +70,9 @@ CONFIG_SCHEMA = voluptuous.Schema(
 
 
 async def async_setup(hass, config):
-    for coreConfig in config[DOMAIN][TAPHOME_CORES]:
+    hass.data[TAPHOME_LANGUAGE] = config[DOMAIN][CONF_LANGUAGE]
+
+    for coreConfig in config[DOMAIN][CONF_CORES]:
         token = coreConfig[CONF_TOKEN]
         tapHomeHttpClientFactory = TapHomeHttpClientFactory()
         tapHomeHttpClient = tapHomeHttpClientFactory.create(token)
@@ -55,40 +81,28 @@ async def async_setup(hass, config):
 
         devices = await tapHomeApiService.async_discovery_devices()
 
-        lightIds = coreConfig[CONF_LIGHTS]
-        if lightIds:
-            lights = filter_devices_by_ids(devices, lightIds)
-            platformConfig = create_platform_config(tapHomeApiService, lights)
+        platforms = [
+            {"config": CONF_LIGHTS, "platform": "light"},
+            {"config": CONF_COVERS, "platform": "cover"},
+            {"config": CONF_CLIMATES, "platform": "climate"},
+            {"config": CONF_SWITCHES, "platform": "switch"},
+            {"config": CONF_SENSORS, "platform": "sensor"},
+            {"config": CONF_BINARY_SENSORS, "platform": "binary_sensor"},
+        ]
 
-            hass.async_create_task(
-                async_load_platform(hass, "light", DOMAIN, platformConfig, config)
-            )
+        for platform in platforms:
+            platformDeviceIds = coreConfig[platform["config"]]
+            if platformDeviceIds:
+                platformDevices = filter_devices_by_ids(devices, platformDeviceIds)
+                platformConfig = create_platform_config(
+                    tapHomeApiService, platformDevices
+                )
 
-        coverIds = coreConfig[CONF_COVERS]
-        if coverIds:
-            covers = filter_devices_by_ids(devices, coverIds)
-            platformConfig = create_platform_config(tapHomeApiService, covers)
-
-            hass.async_create_task(
-                async_load_platform(hass, "cover", DOMAIN, platformConfig, config)
-            )
-
-        climateIds = coreConfig[CONF_CLIMATES]
-        if climateIds:
-            climates = filter_devices_by_ids(devices, climateIds)
-
-            hass.async_create_task(
-                async_load_platform(hass, "climate", DOMAIN, climates, config)
-            )
-            
-        switchIds = coreConfig[CONF_SWITCHES]
-        if switchIds:
-            switches = filter_devices_by_ids(devices, switchIds)
-            platformConfig = create_platform_config(tapHomeApiService, switches)
-
-            hass.async_create_task(
-                async_load_platform(hass, "switch", DOMAIN, platformConfig, config)
-            )
+                hass.async_create_task(
+                    async_load_platform(
+                        hass, platform["platform"], DOMAIN, platformConfig, config
+                    )
+                )
 
     return True