+++ taphome_sdk/ValueType.py
@@ -28,7 +28,9 @@ class ValueType(Enum):
     FloodState = 47
     SwitchState = 48
     MultiValueSwitchState = 49
+    Motion = 51
     ButtonPressed = 52
+    Voc = 54
     AnalogInputValue = 55
     TotalImpulseCount = 56
     CurrentHourImpulseCount = 57