+++ changelog.md
@@ -7,3 +7,7 @@
 - Multiple Core unit support
 ## 0.4
 - Climates - thermostats
+## 0.5
+- Switches - power outlet, digital out
+## 0.6
+- Sensors - Humidity, Temperature, Variable, Motion, Generic reed contact, Electric counter (consumption, demand), Brightness, Co2, Wind speed, Pulse counter (total impulse, current hour impulse, frequency)